Weighted Simple Recommendation for the new users
I guess, initially just to show the new movies may not be a good option. For which, we can follow the weighted recommendation based on the number of user votes and the average rating on IMDB.

Weighted Rating (WR) = (v/(v+m).R)+(m/(v+m).C)                  --> which is used for the top rating 250 movies in IMDB

Where,
where,

    v is the number of votes for the movie;
    m is the minimum votes required to be listed in the chart;
    R is the average rating of the movie; And
    C is the mean vote across the whole report

I already generated the output in the simple_recommendation.sql.  When the new user logs into the system, we can show these movie recommendations to the new user.